+++ update.js
@@ -6,8 +6,9 @@ function htmlMaker(text) {
 }
 
 var nouveautes = `
-Nouvelle version de Weda-Helper !
+# [2.4] - 2024-
 
+TODO
 
 
 `
@@ -17,19 +18,17 @@ nouveautes = htmlMaker(nouveautes)
 
 
 var firstStartMessage = `
-Bienvenue sur Weda-Helper !
+<h1> 👋 Bienvenue sur Weda-Helper !</h1>
 
-Pour commencer, vous devez configurer l'extension. Pour cela, cliquez sur l'icône puzzle en haut à droite de votre navigateur.
+👌 Tout devrait fonctionner de base sans configuration, mais vous pouvez personnaliser l'extension dans les options (clic droit sur l'icone W de l'extension puis options).
 
-Je vous conseille de la mettre en favori en cliquant sur la punaise pour la garder visible.
+📌 Je vous conseille de la mettre en favori en cliquant sur la punaise pour la garder visible.
 
-Pour qu'elle fonctionne au mieux :
-- allez dans les options (bouton de droite sur l'icone de l'extension puis option), vérfiez vos choix puis sauvegardez.
-- définissez les raccourcis clavier dans chrome (idem mais cliquez sur gérer les extensions)
+🔧 Je vous encourage également à installer le Companion <a href="https://github.com/Refhi/Weda-Helper-Companion/releases/latest/download/Weda.Companion.exe" target="_blank">disponible ici pour windows</a> et <a href="https://github.com/Refhi/Weda-Helper-Companion/releases/latest/download/Weda.Companion.dmg" target="_blank">ou pour mac</a> pour profiter de fonctionnalités supplémentaires (Impression totale, lien avec le TPE et upload automatisé).
 
-Vous pouvez aussi relire <a href="https://github.com/Refhi/Weda-Helper/" target="_blank">Weda-Helper sur gitHub</a> pour plus de précisions, et y faire des suggestions ou des signalements de bugs. 
+📝 Vous pouvez aussi relire <a href="https://github.com/Refhi/Weda-Helper/" target="_blank">Weda-Helper sur gitHub</a> pour plus de précisions, et y faire des suggestions ou des signalements de bugs. 
 
-Et bien sûr m'encourager sur le <a href="https://communaute.weda.fr/t5/Entraide-Logiciel-Weda/Weda-Helper-et-Weda-Helper-Companion/m-p/2998" target="_blank">Site de la communauté de weda</a>
+💖 Et bien sûr m'encourager sur le <a href="https://communaute.weda.fr/t5/Entraide-Logiciel-Weda/Weda-Helper-et-Weda-Helper-Companion/m-p/2998" target="_blank">Site de la communauté de weda</a>
 
 💰 Si vous le souhaitez vous pouvez également participer à mes frais de développement (écran, abonnement copilot, etc.) via <a href="https://www.paypal.com/paypalme/refhi" target="_blank">Paypal</a> ("entre proches")
 
@@ -39,30 +38,32 @@ Bon courage,
 
 Le dev de Weda-Helper
 
-P.S. Pour aller plus loin n'oubliez pas de voir les fonction du <a href="https://github.com/Refhi/Weda-Helper-Companion/" target="_blank">Companion</a>`;
+P.S. 🔍 Pour aller plus loin n'oubliez pas de voir les fonction du <a href="https://github.com/Refhi/Weda-Helper-Companion/" target="_blank">Companion</a>`;
 firstStartMessage = htmlMaker(firstStartMessage)
 
 
 var updateMessage = `
-Bonjour !
+<strong>👋 Bonjour !</strong>
 
-Weda-Helper vient d'être mis à jour en version ${currentVersion} !
+<strong>✨ Weda-Helper vient d'être mis à jour en version ${currentVersion} !
 
-Je vous conseille d'aller faire un tour dans les options pour vérifier les nouveaux paramètres : bouton de droite sur l'icone de l'extension puis option.
+🔧 Je vous conseille d'aller faire un tour dans les options pour vérifier les nouveaux paramètres : bouton de droite sur l'icone de l'extension puis option.
 
-Voici les nouveautés et les améliorations :
+🚀 Si vous ne l'avez pas encore, n'hésitez pas à tester le Companion : <a href="https://github.com/Refhi/Weda-Helper-Companion/releases/latest/download/Weda.Companion.exe" target="_blank">disponible ici pour windows</a> et <a href="https://github.com/Refhi/Weda-Helper-Companion/releases/latest/download/Weda.Companion.dmg" target="_blank">ou pour mac</a> pour profiter de fonctionnalités supplémentaires (Impression totale, lien avec le TPE et upload automatisé).</strong>
+
+<strong>🌟 Voici les nouveautés et les améliorations :</strong>
 ${nouveautes}
 
 
-Les suggestions et les rapports de bug c'est toujours par là : <a href="https://github.com/Refhi/Weda-Helper/" target="_blank">Weda-Helper sur gitHub</a>
+📝 Les suggestions et les rapports de bug c'est toujours par là : <a href="https://github.com/Refhi/Weda-Helper/" target="_blank">Weda-Helper sur gitHub</a>
 
-Et les encouragements toujours par ici :-)  <a href="https://communaute.weda.fr/t5/Entraide-Logiciel-Weda/Weda-Helper-et-Weda-Helper-Companion/m-p/2998" target="_blank">Site de la communauté de weda</a>
+💖 Et les encouragements toujours par ici :-)  <a href="https://communaute.weda.fr/t5/Entraide-Logiciel-Weda/Weda-Helper-et-Weda-Helper-Companion/m-p/2998" target="_blank">Site de la communauté de weda</a>
 
 <span style="font-size: 3em;">💰</span> Si vous le souhaitez vous pouvez également participer à mes frais de développement (écran, abonnement copilot, etc.) via <a href="https://www.paypal.com/paypalme/refhi" target="_blank">Paypal</a> ("entre proches")
 
 Bon courage,
 
-Le dev de Weda-Helper
+Les devs de Weda-Helper
 `;
 
 updateMessage = htmlMaker(updateMessage)