+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Toutes les modifications notables de ce projet seront documentées dans ce fichi
 ## fix :
 - le focus reste dans le champ de recherche après l'ouverture automatique des prescriptions-types
 - blocage du historyToLeft si une fenêtre de prévisualisation est ouverte pour éviter des superpositions
+- alt+V fonctionne désormais aussi pour valider l'import de documents
 
 # [2.3] - 2024-05-28
 ## refactory :